backend/auth_service/requirements.txt
fastapi[all]
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
python-dotenv
alembic

